Fix inline generation bug #103 (#105)
A bug was introduced somewhere that was causing top-level schema definitions to be generated as part of the inlined schema generation logic. This had resulted in complex polymorphic classes getting over-written with simpler inline-derived versions of those classes.
This PR simply removes the logic that generates referenced object definitions during inline generation. It relies on the fact that these schemas will undergo regular generation.