Move additionalProperties map out of constructor
Previously, fabrikt would generate a `properties` constructor argument
when `additionalProperties` is `true` for an object schema. This would
cause construction of such models when deserializing as jackson has no
way of setting values in that map.

This patch moves the map definition out of the constructor into a
member, allowing the model to be constructed properly. Additional
properties get inserted into and queried from the list using the already
existing `@JsonAnySetter` and `@JsonAnyGetter` functions.