Introduce typed decorator for ApplicationCall
For routes that respond with content, a decorator is passed to controller functions and enables the implementation to use `call.respondTyped()` which enforces the correct return type for the route.

All other `call.respondX` functions are still available and thus the decoration functions as a drop-in replacement for ApplicationCall.

For routes that do not respond with content the regular ApplicationCall is passed.