Kotlin serialization: Numbers (#364)
With Jackson we use Java's BigDecimal for numbers without a format keyword, but because Java types are not serializable with Kotlin Serialization. 

In order to not loose precision we annotate the field with @Contexual and leave it to the user to provide a custom serializer. This is aligns with the recommended approach in the documentation.