add springboot's http interface generator (#383)
SpringBoot's HTTP Interface is a declarative http client like Retrofit or OpenFeign.
https://docs.spring.io/spring-framework/reference/integration/rest-clients.html#rest-http-interface

Since SpringBoot has now removed OpenFeign from the official project and seems to be focusing on this HTTP Interface, we thought it reasonable to add support for it and implemented it.

This client supports ResponseEntity and suspend methods, so we move the relevant code to ClientGeneratorUtils and so on.